int -> double

+++ matriks.java
@@ -1,11 +1,11 @@
 class matriks {
 
-    int [][] Mat = new int[105][105];
+    Double [][] Mat = new Double[105][105];
 
     matriks() {
         for (int i = 1; i <= 100; i++) {
             for (int j = 1 ; j <= 100; j++) {
-                this.Mat[i][j] = 0;
+                this.Mat[i][j] = 0.0;
             }
         }
     }
